+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add additional help text to "Add planting" form.
+
 ## [3.0.0-alpha4] 2026-02-14
 
 ### Added